bs:050214, HelperGenerateChatMessage added
ChatLog changed the method that was doing the filtering. Added new
handler for changing that message and disabled old handles. Left in for
legacy content for the moment. Have yet to relook a output changes you
suggest Togglebutton. Will look at them tonight after dinner. Opened in
game, appears to work as it was last week now. Uploading this patched
version to Curse for folks to use for the time being.