bs:050314, Revert gsub changes due to formatting issue.
Unfortunately, I discovered last night that we can not simple begin a
newline with our modifications because the chat line is already opened.
It's the reason they append our message with AppendText instead of
AddLine. The Helper opens the line, adds time stamp, sender and the
prompt character, before passing down to the part where our message
actually gets processed. Trying to keep my modifications of Carbine code
to a minimum, I decided to return to the method from the previous submit
and cleaned up the code.