bs:051014, More Work, AltParseForContext
Have this parsing for ooc as well as quotes and emotes now. Did some
bandaid patching to DumptoChat to get it to send out. Needs clean up,
but working. Next going to build an options dialog and put in some
switches in the code to turn things on and off. Maybe a color picker or
field for setting colors with hex.