wip 10022015 1920
Fixed the problem. It was an error in the RestoreChatLogSettings
function. The self.tViewed table is initialized in the metatable, but a
empty table does not register as a nil, so it was running all the time
and overwriting tData on the windows with nil. Problem exists in 1-5-18,
will effect any character that's brand new and has never used Killroy.